bigquery_adk_agent/requirements.txt
google-cloud-bigquery
google-adk
google-cloud-storage
python-dotenv
sqlparse
matplotlib
pandas
db-dtypes
